# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: RoundPenny API
  version: "1.0.0"
  description: RoundPenny — Round-up micro-transaction platform API
servers:
- url: http://localhost:80
  description: Kong Gateway (local development)
paths:
...
  /v1/fees/config:
    get:
      summary: Get fee configuration
      tags: [Fee]
      responses:
        "200":
          description: Fee configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeConfig"
        "500": { description: Internal error }
  /v1/investments/portfolios:
    get:
      summary: Get investment portfolios
      tags: [Investment]
      responses:
        "200":
          description: Portfolio data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
        "500": { description: Internal error }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FeeConfig:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        percentage: { type: number }
        flat_fee: { type: number }
        min_fee: { type: number }
        max_fee: { type: number }
        currency: { type: string }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Portfolio:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: string }
        total_value: { type: number }
        cash_balance: { type: number }
        currency: { type: string }
        holdings:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioHolding"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortfolioHolding:
      type: object
      properties:
        id: { type: string, format: uuid }
        portfolio_id: { type: string, format: uuid }
        asset: { type: string }
        quantity: { type: number }
        purchase_price: { type: number }
        current_price: { type: number }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...